signal_migration/requirements.txt
pandas==1.2.5
requests==2.25.1
boto3==1.17.102
pyarrow==4.0.1
python-dotenv==0.18.0